requirements.txt
connexion==2.14.1
Flask==2.2.2
Flask-Cors==3.0.10
flask-rebar==2.2.1
Jinja2==3.1.2
numpy==1.22.0
PyYAML==5.4.1
transformers==4.25.1
